Reduce code in Modal code path for backups
Use Modal's native retry handling

Retry on Modal ConnectionError

Break out async Modal executors into separate modules
