+++ .coveragerc
@@ -3,4 +3,4 @@ omit =
     */tests/*
     cubed/array_api/*
     cubed/extensions/*
-    cubed/runtime/executors/modal.py
+    cubed/runtime/executors/modal*.py
