+++ examples/modal-add-random.py
@@ -5,7 +5,7 @@ import cubed.array_api as xp
 import cubed.random
 from cubed.extensions.history import HistoryCallback
 from cubed.extensions.timeline import TimelineVisualizationCallback
-from cubed.runtime.executors.modal import AsyncModalDagExecutor
+from cubed.runtime.executors.modal_async import AsyncModalDagExecutor
 
 if __name__ == "__main__":
     tmp_path = sys.argv[1]
